4/23/18
I added the cell builder and made it so when you press the button, it launches the building process of all the cells. The array of boxes is in the main and can access which cell it is in, what row and col it is in, and the value of the number stored in the textfield at the index location.