4/27/18
I made it work so that if there is only one possibility for the box, it'll fill it in and work around it to solve the rest of it but now I'm working on solving it by using the other boxes around it. I created a cell class that with gather the intersections in each of the rows and columns of the cell. This way if one number can go in two or three spaces and it happens to be in the same row or column, then you can say it will be there and it will make sure when you are solving for the possibilities, it will narrow it down.